requirements.txt
aniso8601==9.0.1
boto3==1.24.81
botocore==1.27.81
click==8.1.3
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
Pillow==9.2.0
platformdirs==2.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2
zipp==3.8.1
